anim mod 作成分離
