+++ Assets/HhotateA/AvatarModifyTool/Editor/AnimatorModifier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 59cac6b32b384004da8f72401e0bfc81
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
